+++ spec/rubocop/cli_spec.rb
@@ -984,17 +984,21 @@ RSpec.describe RuboCop::CLI, :isolated_environment do
       create_file('file.rb', 'x=0') # Included by default
       create_file('example', 'x=0')
       create_file('regexp', 'x=0')
+      create_file('vendor/bundle/ruby/2.4.0/gems/backports-3.6.8/.irbrc', 'x=0')
       create_file('.dot1/file.rb', 'x=0') # Hidden but explicitly included
       create_file('.dot2/file.rb', 'x=0') # Hidden, excluded by default
       create_file('.dot3/file.rake', 'x=0') # Hidden, not included by wildcard
       create_file('.rubocop.yml', <<-YAML.strip_indent)
         AllCops:
           Include:
+            - "**/.irbrc"
             - example
             - "**/*.rb"
             - "**/*.rake"
             - !ruby/regexp /regexp$/
             - .dot1/**/*
+          Exclude:
+            - vendor/bundle/**/*
       YAML
       expect(cli.run(%w[--format files])).to eq(1)
       expect($stderr.string).to eq('')